+++ tests/testthat/test_GIRF.R
@@ -17,6 +17,7 @@ mod222s <- GMVAR(data, p=2, M=2, params=params222s, structural_pars=list(W=W_222
 girf1 <- GIRF(mod222s, N=2, R2=2, R1=2, seeds=1:2)
 girf2 <- GIRF(mod222s, variables=2, shock_size=1, N=1, R2=1, R1=1, seeds=1,
               include_mixweights=FALSE, init_values=mod222s$data, ci=0.1)
+girf3 <- GIRF(mod222s, N=2, R2=1, R1=1, which_cumulative=1:2, seeds=1)
 
 test_that("GIRF works correctly", {
   expect_equal(unname(girf1$girf_res[[1]]$point_est[3,]), c(2.7647446, 2.6555342, -0.2440685, 0.2440685), tolerance=1e-4)
@@ -27,4 +28,9 @@ test_that("GIRF works correctly", {
   expect_equal(unname(girf2$girf_res[[1]]$point_est[1,]), c(0.1317841, -0.8456145), tolerance=1e-4)
   expect_equal(unname(girf2$girf_res[[1]]$point_est[2,]), c(-3.333572, -1.151782), tolerance=1e-4)
   expect_equal(unname(girf2$girf_res[[1]]$conf_ints[2, 2, ]), c(-3.333572, -1.151782), tolerance=1e-4)
+
+  expect_equal(unname(girf3$girf_res[[1]]$point_est[,1]), c(1.643883, 3.926812, 6.325788), tolerance=1e-4)
+  expect_equal(unname(girf3$girf_res[[1]]$point_est[,2]), c(1.307096, 3.234228, 5.572740), tolerance=1e-4)
+  expect_equal(unname(girf3$girf_res[[1]]$conf_ints[1, 2, ]), c(1.643883, 1.307096, 0.000000, 0.000000), tolerance=1e-4)
+
 })