+++ NEWS.md
@@ -66,3 +66,4 @@
 
 * Updated some of the examples, readme, and vignette.
 * Fixed a bug that caused error in estimation of GIRF with very large shock size.
+* Fixed a bug that caused error in estimation of AR constrained models when initial population is used in the genetic algorithm.